Constify, isolate phoneme list/table sizes
Also correct some minor code errors; no code changes but in a couple of
places # directives were missing or misplaced.  Mainly this adds 'const'
to paramter declarations on functions and variables within functions.
Some #define counts of array sizes have been moved from header files,
this causes a change in tests/readclause.c which is the only thing
outside translate.c using N_TR_SOURCE; readclause.c now uses a
different (larger) number.

This has been tested on gcc-8.4.0 with -O2 (the default) and -O0 -g.  In
addition -Wall and -Wextra were turned on to enable all the common
warnings.  The compilation is performed additionally with -Werror so
that it fails if there are any warnings.  This requires the following to
be disabled:

FEXTRA='-Werror -Wno-implicit-fallthrough -Wno-format-overflow -Wno-format-truncation -Wno-sign-compare -Wno-format -Wno-int-conversion'

Makefile.am supports $(FEXTRA) to add those flags to the compilation at
build time; -Werror causes configure to fail if set in CFLAGS.

The 'int-conversion', 'format' and 'sign-compare' errors should probably
be examined.  The 'format-overflow' and 'format-trunction' could be
fixed (potential memory overwrite).  'implicit-fallthrough' should
probably just be ignored; the code uses fall through on switch cases
ubiquitously.

+++ src/espeak-ng.c
@@ -334,7 +334,7 @@ int main(int argc, char **argv)
 
 	int option_index = 0;
 	int c;
-	int ix;
+	size_t ix;
 	char *optarg2;
 	int value;
 	int flag_stdin = 0;
